* merge done until commit '7313587ea0c01e3e9fd8044903f3bd0fb0a986fc':
  Fly bonus added
  PV+, Mega, SlowSpeed bonus added, bug fixes

+++ runner/header/Model.h
@@ -53,7 +53,7 @@ public:
 
 protected:
     //=== ATTRIBUTES
-    unsigned int m_width, m_height;
+    float m_width, m_height;
     const std::chrono::system_clock::time_point m_programBeginningTime;
 
 private:
